+++ Runtime/UIToolkitAttribute.cs
@@ -1,0 +1,13 @@
+﻿#if UNITY_2022_2_OR_NEWER && !SAINTSFIELD_UI_TOOLKIT_DISABLE
+using UnityEngine;
+
+namespace SaintsField
+{
+
+    public class UIToolkitAttribute: PropertyAttribute, ISaintsAttribute
+    {
+        public SaintsAttributeType AttributeType => SaintsAttributeType.Other;
+        public string GroupBy => "";
+    }
+}
+#endif