+++ Samples~/Scripts/AddComponentExample.cs
@@ -6,5 +6,7 @@ namespace SaintsField.Samples.Scripts
     {
         [AddComponent, GetComponent] public Dummy myDummy;
         [AddComponent(typeof(BoxCollider)), GetComponent] public GameObject myObj;
+        [UIToolkit]
+        public string fallbackStyle;
     }
 }