+++ Samples~/Scripts/UIKit/UIKitExample2.cs
@@ -1,4 +1,5 @@
 using System;
+using SaintsField.Unsaintly;
 using UnityEngine;
 // using NaughtyAttributes;
 
@@ -10,7 +11,25 @@ namespace SaintsField.Samples.Scripts.UIKit
         // [ShowNativeProperty] public Color blue => Color.blue;
 
         // public int normal;
-        [UIKitPropDec] public string content;
+        public string content;
+        [UIToolkit] public string contentFixed;
+
+        [field: SerializeField] public string AutoContent { get; private set; }
+        [field: SerializeField, UIToolkit] public string AutoContentFixed { get; private set; }
+
+        [ShowInInspector, Ordered] public const float MyConstFloat = 3.14f;
+        [ShowInInspector, Ordered] public static readonly Color MyColor = Color.green;
+
+        [ShowInInspector, Ordered]
+        public Color AutoColor
+        {
+            get => Color.green;
+            // ReSharper disable once ValueParameterNotUsed
+            set
+            {
+                // nothing
+            }
+        }
 
         // [Serializable]
         // public struct MyStruct