+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 2.0.2 ##
+
+1.  When using UI Toolkit mixed with default field with no drawer, now you can add `[UIToolkit]` so `SaintsField` will try to align the label width the same way as UI Toolkit does.
+2.  `UnsaintlyEditor` will try to fix the label width issue when using UI Toolkit. If you're using `UnsaintlyEditor`, you do not need the `[UIToolkit]` attribute.
+
+UI Toolkit supports are experimental, you can disable it by adding a custom marco `SAINTSFIELD_UI_TOOLKIT_DISABLE`
+
 ## 2.0.1 ##
 
 Fix UI Toolkit breaks old Unity versions.